% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{methods}
\name{predict.dblr}
\alias{predict.dblr}
\title{Discrete Boosting Logistic Regression Prediction}
\usage{
\method{predict}{dblr}(object, newdata, type = "response", ...)
}
\arguments{
\item{object}{A fitted dblr model object, which should be returned by calling dblr_train function}

\item{newdata}{A data.frame contains the samples to predict}

\item{type}{Control the output of prediction. Default: 'response' means probability; 'Link' would produce the linear part; 'mapped' would produce a data.frame filling with the coefficients of the model}

\item{...}{further arguments passed to or from other methods}
}
\value{
Returns a vector of prediction or a data.frame
}
\description{
predict.dblr makes predictions on new data set given the fitted dblr model object.
}
